+++ src/main/java/subway/ui/LineController.java
@@ -1,16 +1,21 @@
 package subway.ui;
 
+import java.net.URI;
+import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import subway.application.LineService;
 import subway.application.RouteService;
 import subway.dto.LineRequest;
 import subway.dto.LineResponse;
 import subway.dto.StationsByLineResponse;
-
-import java.net.URI;
-import java.sql.SQLException;
-import java.util.List;
 
 @RestController
 @RequestMapping("/lines")
@@ -52,18 +57,4 @@ public class LineController {
         return ResponseEntity.noContent().build();
     }
 
-    @ExceptionHandler(SQLException.class)
-    public ResponseEntity<Void> handleSQLException() {
-        return ResponseEntity.badRequest().build();
-    }
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<String> handleIllegalArgumentException(final Exception e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
-    
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<String> handleException(final Exception e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
 }
